LayoutTests/inspector/timeline/timeline-time-stamp-expected.txt

Tests the Timeline API timeStamp feature

TimeStamp Properties:
{
    startTime : <number>
    data : {
        message : "TIME STAMP"
    }
    type : "TimeStamp"
    usedHeapSize : <number>
    totalHeapSize : <number>
}
